yaml int thing fix

+++ chunk_fdmt_fb2bin.py
@@ -292,7 +292,7 @@ logging.debug(f"Only writing {maxDT}:-{maxDT} slice in time, should be {out.shap
 for ii in fouts_indices:
     fouts[ii].write(out[dm_slices[ii], maxDT:-maxDT].ravel())
 
-unravel_plan = [out.shape[1] - 2*maxDT]
+unravel_plan = [int(out.shape[1] - 2*maxDT)]
 
 t2 = time.perf_counter()
 logging.info(f"Completed gulp 0 in {t1-t0} s, wrote in {t2-t1} s\n")
@@ -314,8 +314,8 @@ if ngulps > 1:
         for ii in fouts_indices:
             fouts[ii].write(prev_arr[dm_slices[ii],:].ravel())
             fouts[ii].write(out[dm_slices[ii], maxDT:-maxDT].ravel())
-        unravel_plan.append(maxDT)
-        unravel_plan.append(out.shape[1] - 2*maxDT)
+        unravel_plan.append(int(maxDT))
+        unravel_plan.append(int(out.shape[1] - 2*maxDT))
         logging.debug(f"Completed gulp {g}")
 
         # reset for next gulp
